clean up jest config and testing stuff
