+++ package.json
@@ -14,7 +14,7 @@
     "storybook": "turbo run sb --filter=storybook... --parallel",
     "lint": "turbo run lint",
     "format": "prettier --write \"**/*.{js,ts,tsx,md}\"",
-    "test": "yarn workspace @eden/package-ui test",
+    "test": "turbo run test -- --passWithNoTests",
     "test:coverage": "yarn workspace @eden/package-ui test --coverage",
     "test:watch": "yarn workspace @eden/package-ui test --watch",
     "test:turbo": "turbo run test --filter=@eden/package-ui",
@@ -41,7 +41,17 @@
     "npm": ">=7.0.0",
     "node": ">=14.0.0"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@swc/core": "^1.3.23",
+    "@swc/jest": "^0.2.24",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.4",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1"
+  },
   "packageManager": "yarn@1.22.18",
   "lint-staged": {
     "*": "prettier --write --ignore-unknown"
