--- packages/config/jest/jest-preset.js
@@ -1,13 +1,0 @@
-module.exports = {
-  roots: ["<rootDir>", "<rootDir>/packages/ui"],
-  transform: {
-    "^.+\\.tsx?$": "ts-jest",
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
-  modulePathIgnorePatterns: [
-    "<rootDir>/test/__fixtures__",
-    "<rootDir>/node_modules",
-    "<rootDir>/dist",
-  ],
-  preset: "ts-jest",
-};
