+++ packages/config/package.json
@@ -4,8 +4,7 @@
   "main": "index.js",
   "private": true,
   "files": [
-    "eslint-preset.js",
-    "./jest/jest-preset.js"
+    "eslint-preset.js"
   ],
   "scripts": {
     "clean": "rm -rf node_modules",
@@ -16,14 +15,8 @@
   },
   "devDependencies": {
     "@eden/package-tsconfig": "*",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "13.3.0",
-    "@types/jest": "^28.1.6",
     "eslint": "^8.24.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^28.1.2",
-    "tailwind-scrollbar-hide": "^1.1.7",
-    "ts-jest": "^28.0.7"
+    "tailwind-scrollbar-hide": "^1.1.7"
   }
 }
