+++ packages/jest-presets/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "jest-presets",
+  "private": true,
+  "version": "0.0.0"
+}
