--- packages/ui/.swcrc
@@ -1,20 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/swcrc",
-  "jsc": {
-    "parser": {
-      "syntax": "typescript",
-      "dynamicImport": false,
-      "decorators": false,
-      "tsx": true
-    },
-    "target": "es5",
-    "loose": false,
-    "externalHelpers": false,
-    "transform": {
-      "react": {
-        "runtime": "automatic"
-      }
-    }
-  },
-  "minify": false
-}
