--- packages/ui/jest.config.js
@@ -1,41 +1,0 @@
-// module.exports = {
-//   projects: ["<rootDir>/packages/*/jest.config.js"],
-// };
-
-const config = {
-  testMatch: ["**/?(*.)(spec|test).{js,jsx,ts,tsx}"],
-  coveragePathIgnorePatterns: [
-    ".turbo/",
-    "./coverage/",
-    "./node_modules/",
-    "./src/index.ts",
-    "./src/styles.css",
-    "./src/elements/index.ts",
-    "./src/components/index.ts",
-    "./src/components/*",
-    "./utils",
-    "./types",
-    ".eslintrc.js",
-    "jest.config.js",
-    "tailwind.config.js",
-  ],
-  collectCoverageFrom: ["**/*.{js,ts,tsx,jsx}", "!**/*.stories.*"],
-  moduleNameMapper: {
-    "\\.s?css$": "identity-obj-proxy",
-  },
-  coverageThreshold: {
-    global: {
-      branches: 50,
-      functions: 50,
-      lines: 50,
-      statements: 50,
-    },
-  },
-  setupFilesAfterEnv: ["../config/jest/setup.ts"],
-  testEnvironment: "jsdom",
-  transform: {
-    "^.+\\.(t|j)sx?$": ["@swc/jest"],
-  },
-};
-
-module.exports = config;
