+++ packages/ui/jest.config.ts
@@ -1,0 +1,7 @@
+import type { Config } from "jest";
+
+const config: Config = {
+  preset: "jest-presets/jest/jsdom",
+};
+
+export default config;
