+++ packages/ui/package.json
@@ -19,8 +19,6 @@
     "@headlessui/react": "^1.6.6",
     "@heroicons/react": "^1.0.6",
     "@jaames/iro": "^5.5.2",
-    "@swc/core": "^1.3.21",
-    "@swc/jest": "^0.2.23",
     "@testing-library/user-event": "^14.4.3",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.8",
@@ -30,6 +28,7 @@
     "emoji-picker-react": "^3.6.1",
     "eslint": "^8.20.0",
     "identity-obj-proxy": "^3.0.0",
+    "jest-presets": "*",
     "next-router-mock": "^0.8.0",
     "react-chartjs-2": "^4.3.1",
     "react-confetti": "^6.1.0",
